# internal/repo/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.27.0

package db

import (
	"database/sql"
	"time"
)

type File struct {
	Uuid      string
	Url       string
	Notes     string
	Extension string
	Pit       time.Time
}

type Permalink struct {
	Uuid            string
	FileUuid        string
	DurationSeconds int64
	Expires         time.Time
	Pit             time.Time
}

type Picture struct {
	ID          int64
	Author      string
	Url         string
	Description string
	Extension   string
	NumLikes    int64
	NumDislikes int64
	Pit         time.Time
}

type SurveyState struct {
	ID   int64
	Data []byte
	Pit  time.Time
}

type Visitor struct {
	ID      int64
	Path    string
	Message string
	Ip      sql.NullString
	City    sql.NullString
	Region  sql.NullString
	Country sql.NullString
	Pit     time.Time
}